+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -106,6 +106,7 @@
     <Compile Include="BlastLayer.cs" />
     <Compile Include="BlastTools.cs" />
     <Compile Include="BlastUnit.cs" />
+    <Compile Include="ComboBoxItem.cs" />
     <Compile Include="CorruptCore.cs" />
     <Compile Include="Diff.cs" />
     <Compile Include="Enums.cs" />
